+++ src/main/kotlin/de/schwarzland/mavenup/service/MavenUpSettings.kt
@@ -74,6 +74,7 @@ class MavenUpSettings : PersistentStateComponent<MavenUpSettings.State> {
      * @property stopAfterCentralSuccess Bestimmt, ob nach einer erfolgreichen Abfrage von Maven Central keine weiteren privaten Repositories abgefragt werden.
      * @property offerAllVersions Bestimmt, ob in der Versionsauswahl alle verfügbaren Versionen (inklusive älterer) angeboten werden, statt nur Versionen `>=` der aktuellen Version.
      * @property confirmVersionReset Bestimmt, ob vor dem Zurücksetzen aller Versionsauswahlen ein Bestätigungsdialog angezeigt wird.
+     * @property autoSearchVersions Bestimmt, ob nach dem automatischen Neuladen der Projektdaten (Tool-Window-Start, Maven-Import/Resync) unmittelbar online nach neuen Versionen gesucht wird.
      * @property addVulnerabilityFixComment Bestimmt, ob beim Anlegen eines gepinnten `dependencyManagement`-Eintrags zur Behebung einer Schwachstelle ein erklärender XML-Kommentar eingefügt wird.
      */
     data class State(
@@ -91,6 +92,7 @@ class MavenUpSettings : PersistentStateComponent<MavenUpSettings.State> {
         var selectLatestMinorVersion: Boolean = false,
         var offerAllVersions: Boolean = false,
         var confirmVersionReset: Boolean = true,
+        var autoSearchVersions: Boolean = true,
         var addVulnerabilityFixComment: Boolean = true
     ) {
         /**